为 Android ARM64 使用固定 39-bit 用户态地址
64 位堆、mmap、栈、SVC、LR 落到真机同档的固定高位区间，/proc/self/maps 按真实映射生成。32 位与 iOS 布局不变。brk 改为返回 long，避免高位截断。默认 Unicorn1 可映射并关闭；Windows Unicorn2 在 39-bit 映射后 uc_close 会崩溃，因此 Android64 的 destroy 不调用 nativeDestroy。Dynarmic/KVM/Hypervisor 仍受 36-bit 页表限制。

+++ unidbg-android/src/main/java/com/github/unidbg/linux/ARM32SyscallHandler.java
@@ -1925,7 +1925,7 @@ public class ARM32SyscallHandler extends AndroidSyscallHandler {
         if (log.isDebugEnabled()) {
             log.debug("brk address=0x{}", Long.toHexString(address));
         }
-        return emulator.getMemory().brk(address);
+        return (int) emulator.getMemory().brk(address);
     }
 
     private int mprotect(Backend backend, Emulator<?> emulator) {
